move ttnn fused tests from apc to l2nightly
