+++ tests/ttnn/nightly/unit_tests/operations/fused/test_batch_norm.py
