+++ tests/ttnn/nightly/unit_tests/operations/fused/test_group_norm.py
