+++ tests/ttnn/nightly/unit_tests/operations/fused/test_rms_norm.py
