+++ tests/ttnn/nightly/unit_tests/operations/fused/test_softmax.py
